Ticket 37190: Show columns in the R export vignette?
